requirements.txt
ollama
numpy 
pandas 
networkx
matplotlib
tqdm
igraph 
ollama 
openai
ges
causal-learn 
lingam 
pyyaml
torch